test-requirements.txt
typing-extensions
mock
coveralls
django-storages
boto3
google-cloud-storage
pytest
pytest-django
moto[all]
tox
